Made improvements and refactored code. Added pypdf and function to load PDFs.
