+++ fervent-api/src/config/constants.py
@@ -1,3 +1,4 @@
 # Constants
 CHROMA_DIR = "tmp/chroma"
 DOCUMENT_DIR = "tmp/sys_files"
+USERS_DIR = "tmp/users"
