+++ fervent-api/src/controllers/vectorstore_controller.py
@@ -1,6 +1,6 @@
 from fastapi.responses import JSONResponse
 
-from src.services.vector_store import init_chromadb, query_chromadb
+from src.services.vectorstores.chroma import init_chromadb, query_chromadb
 
 import traceback
