+++ fervent-api/src/services/llm_chains.py
@@ -1,6 +1,6 @@
 from src.services.llm import get_llm_prompt_response, get_llm_conversational_response
-from src.services.vector_db import query_chromadb
-from src.utils.vector_db_utils import join_similarity_search_results, get_content_sources
+from src.services.vectorstores.chroma import query_chromadb
+from src.utils.vectorstores.vectorstore_utils import join_similarity_search_results, get_content_sources
 from src.utils.exception_utils import ERR_NO_MATCH_FOUND
